Menu Created
the menu page start.html has now two different buttons that will lead to the game and the leaderboard. style.css was created to for the design of the page, with features such as hover on buttons. A background was added too and adjusted to be resposive in the style.css file. An javascript-file was also created to be able to fill the index bodypart with the content of the start.html file.