Leaderboard page created
The first version of the leaderboard is now created. A table is created but with mockup data for now. main.js was updated too for making it possible to go from the menu to the leaderboard page. A backbutton was also added to the leaderboard page, to be able to use this main.js was updated to look for a click on it. A new css file was added to handle the leaderboard page, more specific the table. A link to this as added to the index.html. Next up is to create the game page.