First version of game page
The first version of the game page is now created, an html, css and a js file is all added to support this page. in the html file a canvas is added and a back button like in the leaderboard page. In the game.js 4 functions has been created, loadCanvas is the first method which is called from main.js when the game should load. Then the function start is called which just start the function gameLoop which will run over and over. The last function is gameOver but has no code yet. The game does not work yet, the only thing that happens is that a white rect is printed on the canvas, this is the snake. A red rect is also printed, this is the apple. Some styling for this rect is also done in the gameLoop. The css file was created to help make the cavas resposive, so it should work both on big and small screens. For the next update the snake should move with different keys pressed.