Moving Snake
Divided the gameLoop into more function for easier navigation.
It is now possible to move around the snake on the canvas with up, down, left and right key. The new function keyPressed is the hte function that handels when a key is pressed to move the snake. In the upcoming update the game will end when snake touch itself or the walls, it should also grow when it passes an apple