The game will now end if snake touch itself or the wall
The game will now end if the snake touch itself or touch a wall.
The function wallHanlder were created for wall touch.
SnakeHandler got updated code to detect if snake touch itself.
A variable were also added to be able to know if the game is active or not, this is used when pressing the back button but also in the future to start a new game.
The next update should include for the user to be ablo too choose a username and also to press a key to play again.